Add a simple concurrent object pool.
The pool works on generic types and requires just two functions to be created, one for instantiating new objects and an other to reset old ones. It also optionally supports a dispose method which is called when an object is removed from the pool. The pool also features profiling capabilities, accessible from the "Profile" property. Unit tests included.