Fixes #59
This commit tries to improve performance in two ways. First and
foremost, it highers the number of milliseconds that need to pass before
a unit is forced to try and evade blocks (from 400ms to 1000ms). It also
uses the TightGrid implementation to check for collisions in the
pathfinding instead of doing it O(n)